more TS conversions
