+++ src/WithResponsiveProps.tsx
@@ -4,6 +4,10 @@ import ResponsiveContext from './ResponsiveContext';
 import useThreshold from './useThreshold';
 import defaultThresholdMap from './defaultThresholdMap';
 
+export interface ResponsivePropsConfig {
+  propKeys: Array<string>;
+}
+
 // This component takes a given property, like size, and based on the current threshold replaces that value
 // for example on a mobile phone the value for size would be replaced with the value from xs
 // e.g. size={{xs: 'small', 'md': 'large'}}
@@ -11,8 +15,8 @@ import defaultThresholdMap from './defaultThresholdMap';
 //   propKeys: [ 'size' ],
 // }
 
-const WithResponsiveProps = (configuration) => (WrappedComponent) => {
-  const Component = (props) => {
+const WithResponsiveProps = (configuration: ResponsivePropsConfig) => (WrappedComponent: React.ComponentType) => {
+  const Component = (props: any) => {
     const threshold = useThreshold();
 
     const responsiveContext = useContext(ResponsiveContext);
